requirements.txt
openai-whisper
pyaudio
pynput
pyperclip 